+++ packages/evm-connector/contracts/data-services/MainDemoConsumerBase.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.4;
 import "../core/RedstoneConsumerNumericBase.sol";
 
 contract MainDemoConsumerBase is RedstoneConsumerNumericBase {
-  function getUniqueSignersThreshold() public view virtual override returns (uint256) {
+  function getUniqueSignersThreshold() public view virtual override returns (uint8) {
     return 1;
   }
 
@@ -14,7 +14,7 @@ contract MainDemoConsumerBase is RedstoneConsumerNumericBase {
     view
     virtual
     override
-    returns (uint256)
+    returns (uint8)
   {
     if (_signerAddress == 0x0C39486f770B26F5527BBBf942726537986Cd7eb) {
       return 0;
