+++ packages/evm-connector/contracts/samples/Benchmark.sol
@@ -8,13 +8,13 @@ contract Benchmark is RedstoneConsumerNumericMock {
   // We use this storage variable to avoid annoying compilation
   // warnings about function state mutability
   uint256 someStorageVar = 0;
-  uint256 uniqueSignersThreshold = 1;
+  uint8 uniqueSignersThreshold = 1;
 
-  function getUniqueSignersThreshold() public view virtual override returns (uint256) {
+  function getUniqueSignersThreshold() public view virtual override returns (uint8) {
     return uniqueSignersThreshold;
   }
 
-  function updateUniqueSignersThreshold(uint256 newUniqueSignersThreshold) external {
+  function updateUniqueSignersThreshold(uint8 newUniqueSignersThreshold) external {
     uniqueSignersThreshold = newUniqueSignersThreshold;
   }
 
@@ -38,7 +38,7 @@ contract Benchmark is RedstoneConsumerNumericMock {
     public
     view
     override
-    returns (uint256)
+    returns (uint8)
   {
     return getAllMockAuthorised(_signerAddress);
   }
